app/requirements/development.txt
-r ./common.txt

django-extensions==3.2.1
flake8==6.0.0
pre-commit==3.0.2
pytest==7.2.1
pytest-django==4.5.2